всё в целом работает
