+++ package.json
@@ -13,12 +13,15 @@
   "devDependencies": {
     "browser-sync": "2.26.14",
     "eslint": "7.18.0",
-    "eslint-config-htmlacademy": "2.0.0"
+    "eslint-config-htmlacademy": "2.0.0",
+    "webpack": "5.28.0",
+    "webpack-cli": "4.5.0"
   },
   "license": "MIT",
   "dependencies": {},
   "scripts": {
-    "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
+    "start": "browser-sync start --server \"build/\" --no-ui --files \"source/**/*.js\"",
+    "build": "webpack --mode production",
     "test": "eslint js/"
   },
   "engines": {
